Implemented all existing datapar algorithms using Boost.SIMD
- factoring out code common to both backends
- adding build system support for Boost.SIMD
- unified configurations for Vc and Boost.SIMD
- flyby: cleaning up the Vc backend
