+++ CMakeLists.txt
@@ -243,10 +243,19 @@ hpx_option(HPX_WITH_SYCL BOOL
 ################################################################################
 # HPX datapar configuration
 ################################################################################
-hpx_option(HPX_WITH_VC_DATAPAR BOOL
-  "Enable data parallel algorithm support (default: OFF)" OFF ADVANCED)
-if(HPX_WITH_VC_DATAPAR)
+hpx_option(HPX_WITH_DATAPAR_VC BOOL
+  "Enable data parallel algorithm support using the external Vc library (default: OFF)" OFF ADVANCED)
+hpx_option(HPX_WITH_DATAPAR_BOOST_SIMD BOOL
+  "Enable data parallel algorithm support using the external Boost.SIMD library (default: OFF)" OFF ADVANCED)
+if(HPX_WITH_DATAPAR_VC AND HPX_WITH_DATAPAR_BOOST_SIMD)
+  hpx_error("Please select only one of the supported external vectorization libraries (HPX_WITH_DATAPAR_VC or HPX_WITH_DATAPAR_BOOST_SIMD)")
+endif()
+
+if(HPX_WITH_DATAPAR_VC)
   include(HPX_SetupVc)
+endif()
+if(HPX_WITH_DATAPAR_BOOST_SIMD)
+  include(HPX_SetupBoostSIMD)
 endif()
 
 ################################################################################
