+++ hpx/parallel/datapar/execution_policy_fwd.hpp
@@ -8,7 +8,7 @@
 
 #include <hpx/config.hpp>
 
-#if defined(HPX_HAVE_VC_DATAPAR)
+#if defined(HPX_HAVE_DATAPAR)
 #include <hpx/parallel/config/inline_namespace.hpp>
 
 namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
