+++ hpx/parallel/exception_list.hpp
@@ -101,7 +101,6 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
             static hpx::future<Result> call(hpx::future<Result> f)
             {
                 HPX_ASSERT(f.has_exception());
-
                 // Intel complains if this is not explicitly moved
                 return std::move(f);
             }
@@ -148,7 +147,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
           : handle_exception_task_impl<Result>
         {};
 
-#if defined(HPX_HAVE_VC_DATAPAR)
+#if defined(HPX_HAVE_DATAPAR)
         ///////////////////////////////////////////////////////////////////////
         template <typename Result>
         struct handle_exception_impl<datapar_task_execution_policy, Result>
