+++ hpx/parallel/execution_policy.hpp
@@ -11,7 +11,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/parallel/config/inline_namespace.hpp>
-#if defined(HPX_HAVE_VC_DATAPAR)
+#if defined(HPX_HAVE_DATAPAR)
 #include <hpx/parallel/datapar/execution_policy.hpp>
 #endif
 #include <hpx/parallel/execution_policy_fwd.hpp>
