+++ hpx/parallel/util/partitioner.hpp
@@ -583,7 +583,7 @@ namespace hpx { namespace parallel { namespace util
             }
         };
 
-#if defined(HPX_HAVE_VC_DATAPAR)
+#if defined(HPX_HAVE_DATAPAR)
         template <typename R, typename Result>
         struct partitioner<datapar_task_execution_policy, R, Result,
             parallel::traits::static_partitioner_tag>
